Chore: 발할라 연동 및 c-its 캐싱 설정 수정 - 경로는 캐싱하고 c-its api 데이터는 갱신되도록 수정(c-its api 멈충으로 인한 개발 대기)

+++ backend/src/main/java/kr/io/pacer/core/dto/response/SignalResponse.java
@@ -12,12 +12,12 @@ import lombok.NoArgsConstructor;
 public class SignalResponse {
     private int     itstId;
     private String  name;
-    private Integer ntPdsgRmdrCs;
-    private Integer etPdsgRmdrCs;
-    private Integer stPdsgRmdrCs;
-    private Integer wtPdsgRmdrCs;
-    private Integer nePdsgRmdrCs;
-    private Integer sePdsgRmdrCs;
-    private Integer swPdsgRmdrCs;
-    private Integer nwPdsgRmdrCs;
+    private Double ntPdsgRmdrCs;
+    private Double etPdsgRmdrCs;
+    private Double stPdsgRmdrCs;
+    private Double wtPdsgRmdrCs;
+    private Double nePdsgRmdrCs;
+    private Double sePdsgRmdrCs;
+    private Double swPdsgRmdrCs;
+    private Double nwPdsgRmdrCs;
 }
